debug of valgrind and start condition, still miss the queue

+++ includes/coders.h
@@ -6,7 +6,7 @@
 /*   By: eel-kerc <eel-kerc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/11 08:28:27 by eel-kerc          #+#    #+#             */
-/*   Updated: 2026/06/18 16:39:57 by eel-kerc         ###   ########.fr       */
+/*   Updated: 2026/06/23 11:35:11 by eel-kerc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,14 +17,13 @@
 # include <stdio.h>
 # include <unistd.h>
 
-typedef struct s_dongle t_dongle;
-typedef struct s_global t_global;
-
+typedef struct s_dongle	t_dongle;
+typedef struct s_global	t_global;
 
 typedef struct s_coder
 {
 	unsigned int	id;
-	unsigned int	last_compiled;
+	long long		last_compiled;
 	unsigned int	nb_compiled;
 	t_dongle		*first_dongle;
 	t_dongle		*second_dongle;
