+++ includes/simulation.h
@@ -6,7 +6,7 @@
 /*   By: eel-kerc <eel-kerc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/11 10:34:14 by eel-kerc          #+#    #+#             */
-/*   Updated: 2026/06/18 16:36:36 by eel-kerc         ###   ########.fr       */
+/*   Updated: 2026/06/23 11:29:52 by eel-kerc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 # define SIMULATION_H
 
 # include "../includes/coders.h"
-#include <sys/time.h>
+# include <sys/time.h>
 
-void	*simulation(void *arg);
-void	edf(t_coder *coder, t_dongle *dongle);
-void	fifo(t_coder *coder, t_dongle *dongle);
-unsigned int	get_time(unsigned int start);
+void			*simulation(void *arg);
+void			edf(t_coder *coder, t_dongle *dongle);
+void			fifo(t_coder *coder, t_dongle *dongle);
+long long		get_time(long long start);
 
 #endif
